requirements.txt
torch=2.4.0+cu121
torchvision=0.19.0+cu121
transformers=4.51.3
pillow =10.4.0
tiktoken=0.7.0
accelerate=1.6.0
blobfile=3.0.0
gradio=5.25.0
